% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{BAMdiagnostic}
\alias{BAMdiagnostic}
\title{A constructor to load all the *Histogram.txt files in the input folder and
assign the data "BAMdiagnostic" class}
\usage{
BAMdiagnostic(folder_path = getwd(), recursive = TRUE)
}
\arguments{
\item{folder_path}{Folder containing output of ATLAS "BAMDiagnostics" task.}

\item{recursive}{bool: To ckeck in subfolders to finds that fit the criteria.
Default is TRUE.}
}
\value{
A filtered dataframe with all the data for all the *Histogram.txt files
in the  input folder that match certain criteria such as file format, fragment
length of reads <=1500 and column names within the files
}
\description{
A constructor to load all the *Histogram.txt files in the input folder and
assign the data "BAMdiagnostic" class
}
\examples{
input_folder <- list_example_folder()
data <- BAMdiagnostic(input_folder)
}
