% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{CreateClassRGInfo}
\alias{CreateClassRGInfo}
\title{A constructor to load data in JSON format and  assign it "RGInfo" class}
\usage{
CreateClassRGInfo(RGInfoFile,readGroups)
}
\arguments{
\item{RGInfoFile}{The path to the JSON file containing the data.}

\item{readGroups}{A character vector specifying the read groups to plot from the JSON data.}
}
\value{
The JSON data loaded from the file, filtered based on the specified read groups (if any), and assigned the class "RGInfo".
}
\description{
A constructor to load data in JSON format and  assign it "RGInfo" class
}
\examples{
RGInfoFile <- list_example("test.json")
readGroups_1 <- c("ERR8666961", "ERR8684188")
data_1 <- CreateClassRGInfo(RGInfoFile,readGroups_1)
data_2 <- CreateClassRGInfo(RGInfoFile)

}
\author{
Aparna Pandey
}
