% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{.extractRecalData}
\alias{.extractRecalData}
\title{Extract data for different sequencing error covariates to plot}
\usage{
.extractRecalData(info, mate, covariate, ...)
}
\arguments{
\item{info}{data of class "RGInfo"}

\item{mate}{mates of a paired-end sequencing read. Mate1 - 5' end of the
fragment , Mate2 - 2' end of the fragment.}

\item{covariate}{Sequencing Error Covariate}

\item{...}{Additional Parameters}
}
\value{
Data extracted for the specified covariate for the specified mate
}
\description{
Extract data for different sequencing error covariates to plot
}
\author{
Aparna Pandey
}
