% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{.extractRho}
\alias{.extractRho}
\title{Extract the data for 'rho' covariate}
\usage{
.extractRho(info, covariate)
}
\arguments{
\item{info}{Data}

\item{covariate}{Sequencing Error Covariate 'rho'}
}
\value{
Data for rho covariate and the data type i.e. rho
}
\description{
Extract the data for 'rho' covariate
}
\author{
Aparna Pandey
}
