% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{.extractSeqErrorCovariates}
\alias{.extractSeqErrorCovariates}
\title{Extract the data for specified covariate}
\usage{
.extractSeqErrorCovariates(info, covariate)
}
\arguments{
\item{info}{Data}

\item{covariate}{Sequencing Error Covariate}
}
\value{
Data for the specified covariate and the type of data (empiric vs polynomial)
}
\description{
Extract the data for specified covariate
}
\author{
Aparna Pandey
}
