% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{.getFileNames}
\alias{.getFileNames}
\title{Function to extract unique part from the name of each dataframe from a list of
dataframes}
\usage{
.getFileNames(list_of_dataframes)
}
\arguments{
\item{list_of_dataframes}{A list of dataframes with data to be plotted}
}
\value{
Names unique to each sample.
}
\description{
Function to extract unique part from the name of each dataframe from a list of
dataframes
}
