% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{.merge_and_store_data}
\alias{.merge_and_store_data}
\title{This function merges and stores the same type of data from a list of multiple dataframes}
\usage{
.merge_and_store_data(dataframe_list)
}
\arguments{
\item{dataframe_list}{List of dataframes containing different type of data for different
Samples}
}
\value{
A dataframe that contains data for allReadGroups from all samples for
the same data type
}
\description{
This function merges and stores the same type of data from a list of multiple dataframes
}
