% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMD.R
\name{.plot_one_side}
\alias{.plot_one_side}
\title{To plot data for one mate. can be mate1 or mate2}
\usage{
.plot_one_side(
  info,
  side,
  ylim = ylim,
  legend = legend,
  xlab = xlab,
  ylab = ylab,
  main = main,
  ...
)
}
\arguments{
\item{info}{data of class "RGInfo".}

\item{side}{3' to 5'("GA3") or 5' to 3'("CT3").}

\item{ylim}{y-axis limits of a plot.}

\item{legend}{bool: To print the legend or not.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{Main heading of the for the plot.}

\item{...}{additional parameters for plotting (optional).}
}
\value{
plot of the Post Mortem Damage for the specified side.
}
\description{
To plot data for one mate. can be mate1 or mate2
}
\author{
Aparna Pandey
}
