% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{.plotBAMdiagnosticMultiple}
\alias{.plotBAMdiagnosticMultiple}
\title{Plot data for multiple samples}
\usage{
.plotBAMdiagnosticMultiple(data, title)
}
\arguments{
\item{data}{Data to be plotted}

\item{title}{Title for the plot}
}
\value{
Plots with data for allReadGroups for all the samples
}
\description{
Plot data for multiple samples
}
