% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{.plotBAMdiagnosticSingle}
\alias{.plotBAMdiagnosticSingle}
\title{Plot data for a single Sample}
\usage{
.plotBAMdiagnosticSingle(data, title, log_scale = TRUE, ...)
}
\arguments{
\item{data}{Single data set to be plotted}

\item{title}{Title for each plot}

\item{log_scale}{y-axis scale is log10. At the moment this is the only option.}

\item{...}{Additional parameters (optional)}
}
\value{
Plot for the Input data
}
\description{
Plot data for a single Sample
}
