% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{.plotCovariateOneMate}
\alias{.plotCovariateOneMate}
\title{Plot covariate data from one of the mates in the readgroup. Can be mate1 or mate2.}
\usage{
.plotCovariateOneMate(
  data,
  legend,
  xlim = xlim,
  ylim = ylim,
  inset = inset,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{data}{Sequencing error data for Specified covariate. Nested list containing
x and y values for different readgroups for a single mate e.g., x = data[[1]]$x, y= data[[1]]$y}

\item{legend}{bool: To print the legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{inset}{to indicate margins for legend.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{...}{additional parameters for plotting (optional).}
}
\value{
Plot for specified Sequencing Error covariate.
}
\description{
Plot covariate data from one of the mates in the readgroup. Can be mate1 or mate2.
}
\author{
Aparna Pandey
}
