% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{.plotRhoOneMate}
\alias{.plotRhoOneMate}
\title{To plot data for rho one mate. can be mate1 or mate2}
\usage{
.plotRhoOneMate(
  data,
  legend,
  xlim = xlim,
  ylim = ylim,
  inset,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{data}{sequencing error data for rho}

\item{legend}{bool: To print the legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be set to c(1,12).}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{inset}{to indicate margins for legend.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{...}{additional parameters for plotting (optional).}
}
\value{
Plot for rho parameter for all base transitions
}
\description{
To plot data for rho one mate. can be mate1 or mate2
}
\author{
Aparna Pandey
}
