% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{test data}
\format{
\subsection{JSON format with 58 readGroups}{

Each with mate1 and mate2 with
-post mortem damage information
-quality information
-mapping quality information
-position information
-context information
-fragment length information
-rho information
}
}
\source{
simulated using 'simulate' task from ATLAS pipeline
for more information look at ATLAS documentation at \url{https://atlaswiki.netlify.app/}
}
\usage{
exampleData
}
\description{
Example data of simulated ATLAS run
}
\keyword{datasets}
