% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{list_example_folder}
\alias{list_example_folder}
\title{Title}
\usage{
list_example_folder(path = NULL)
}
\arguments{
\item{path}{The path containing output of ATLAS "BAMDiagnostics" task. If not
provided, the function returns the path to the folder containming available
example files.}
}
\value{
Path containing example files.
}
\description{
Title
}
\examples{
list_example_folder()
}
