% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMD.R
\name{plot_PMD}
\alias{plot_PMD}
\title{Plot Post Mortem Damage}
\usage{
plot_PMD(
  path,
  side = "both",
  rg_no = length(info),
  legend = TRUE,
  xlim = NULL,
  ylim = NULL,
  multiPannel = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  readGroups = c(),
  ...
)
}
\arguments{
\item{path}{path to the data file.}

\item{side}{5(5' to 3') or 3 (3' to 5') or both.}

\item{rg_no}{how many read groups to plot. Default is all read Groups in the data.}

\item{legend}{bool : To print legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{multiPannel}{bool: to plot two plots (one for each mate) on one page or not.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{readGroups}{read Groups to plot. By default data for all read Groups are plotted.}

\item{...}{additional parameters for plotting (optional).}
}
\value{
plots depicting postmortem damage.
}
\description{
Plot Post Mortem Damage
}
\note{
readGroups need to be provided as 'readGroups = c(some vector)' or 'readGroups = some_variable'. See examples below.
}
\examples{
path <- list_example("test.json")
readGroups_1 <- c("ERR8666961", "ERR8684188")
plot_PMD(path,readGroups=readGroups_1)
plot_PMD(path)
plot_PMD(path,3,readGroups=readGroups_1)
plot_PMD(path,3)
plot_PMD(path,5,readGroups=readGroups_1)
plot_PMD(path,5)

}
\author{
Aparna Pandey
}
