% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{plot_SeqError_covariate}
\alias{plot_SeqError_covariate}
\title{Function to plot Sequencing Error Covariates.}
\usage{
plot_SeqError_covariate(
  path,
  covariate,
  mate = c(1, 2),
  legend = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  multiPannel = TRUE,
  readGroups = c(),
  ...
)
}
\arguments{
\item{path}{path to the data file.}

\item{covariate}{Sequencing Error Covariate of interest}

\item{mate}{mate information to assess if sequencing was single or paired-ed. By default
data is parsed to access information for both mates.}

\item{legend}{bool : To print legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{multiPannel}{bool: to plot two plots (one for each mate) on one page or not.}

\item{readGroups}{read Groups to plot. By default data for all read Groups are plotted.}

\item{...}{additional parameters for plotting (optional).}
}
\value{
Plots depicting data for the specified Sequencing error covariate.
}
\description{
Function to plot Sequencing Error Covariates.
}
\examples{
path <- list_example("test.json")
readGroups_1 <- c("ERR8666961", "ERR8684188")
Covariates = list('quality','position','context','fragmentLength','mappingQuality','rho')
for (i in Covariates){
plot_SeqError_covariate(path,i,1);
plot_SeqError_covariate(path,i,2);
plot_SeqError_covariate(path,i);
}

}
\author{
Aparna Pandey
}
