% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{plotBAMdiagnostic.BAMdiagnostic}
\alias{plotBAMdiagnostic.BAMdiagnostic}
\title{Function takes data objects with "BAMdiagnostic" class and plots the data}
\usage{
\method{plotBAMdiagnostic}{BAMdiagnostic}(data, ...)
}
\arguments{
\item{data}{Data to be plotted}

\item{...}{Additional parameters (optional).}
}
\value{
Plots of the input data.
}
\description{
Function takes data objects with "BAMdiagnostic" class and plots the data
}
