% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{plotBAMdiagnostic}
\alias{plotBAMdiagnostic}
\title{Generic function plot data with "BAMdiagnostics" class}
\usage{
plotBAMdiagnostic(data, ...)
}
\arguments{
\item{data}{Data to plot that has "BAMDiagnostics" class.}

\item{...}{additional parameters to pass on to the function.}
}
\description{
Generic function plot data with "BAMdiagnostics" class
}
