% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMD.R
\name{plotPMD}
\alias{plotPMD}
\title{Generic method to plot PMD for objects with class "RGInfo"}
\usage{
plotPMD(obj, ...)
}
\arguments{
\item{obj}{A RGInfo object}

\item{\dots}{Further arguments that can be passed for plotting}
}
\description{
Generic method to plot PMD for objects with class "RGInfo"
}
\author{
Aparna Pandey
}
