% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMD.R
\name{plotPMD.RGInfo}
\alias{plotPMD.RGInfo}
\title{Function to plot Postmortem Damage that takes object with class "RGInfo" and additional parameters as input}
\usage{
\method{plotPMD}{RGInfo}(
  info,
  side = "both",
  rg_no = length(info),
  legend = TRUE,
  xlim = NULL,
  ylim = NULL,
  multiPannel = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{info}{data in JSON format with "RGInfo" class.}

\item{side}{5(5' to 3') or 3 (3' to 5') or both.}

\item{rg_no}{how many read groups to plot. Default is all read Groups in the data.}

\item{legend}{bool : To print legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{multiPannel}{bool: to plot two plots (one for each mate) on one page or not.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{...}{additional parameters for plotting (optional).}
}
\description{
Function to plot Postmortem Damage that takes object with class "RGInfo" and additional parameters as input
}
\author{
Aparna Pandey
}
