% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{plotRecal}
\alias{plotRecal}
\title{Generic method to plot Sequencing Error covariates for objects with class "RGInfo"}
\usage{
plotRecal(obj, ...)
}
\arguments{
\item{obj}{Data of class "RGInfo"}

\item{...}{additional parameters}
}
\description{
Generic method to plot Sequencing Error covariates for objects with class "RGInfo"
}
\author{
Aparna Pandey
}
