% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecal.R
\name{plotRecal.RGInfo}
\alias{plotRecal.RGInfo}
\title{Function to plot Sequencing Error Covariates that takes object with class "RGInfo" and additional parameters as input}
\usage{
\method{plotRecal}{RGInfo}(
  info,
  covariate,
  mate = c(1, 2),
  legend = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  multiPannel = TRUE,
  ...
)
}
\arguments{
\item{info}{Data in JSON format with "RGInfo" class}

\item{covariate}{Sequencing Error Covariate.}

\item{mate}{mate information to assess if sequencing was single or paired-ed. By default
data is parsed to access information for both mates.}

\item{legend}{bool : To print legend or not.}

\item{xlim}{x-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{ylim}{y-axis limits of a plot. Default is Null and will be calculate based on the data if no input is provided.}

\item{xlab}{x-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{ylab}{y-axis label. Default is Null and will be generated based on data labels if no input is provided.}

\item{main}{Main heading of the for the plot. Default is Null and will be generated based on data labels if no input is provided.}

\item{multiPannel}{bool: to plot two plots (one for each mate) on one page or not.}

\item{...}{additional parameters for plotting (optional).}
}
\description{
Function to plot Sequencing Error Covariates that takes object with class "RGInfo" and additional parameters as input
}
\author{
Aparna Pandey
}
