% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAMdiagnosticHistogram.r
\name{plot_BAMdiagnostic}
\alias{plot_BAMdiagnostic}
\title{Function to plot BAMDiagnostic histograms}
\usage{
plot_BAMdiagnostic(
  input_path,
  recursive = TRUE,
  output_directory_name = getwd(),
  ...
)
}
\arguments{
\item{input_path}{Path to the folder with all BAM diagnostic *_histogram.txt files}

\item{recursive}{bool: To plot data from subfolders within the input folder path or not.}

\item{output_directory_name}{: Path to an output directory. Default is the input_path.
A plots folder will be created in the default output directory to store the plots.}

\item{...}{Additional parameters. Optional.}
}
\value{
pdf and png images of the histograms
}
\description{
Function to plot BAMDiagnostic histograms
}
\examples{
input_path <- list_example_folder()
plot_BAMdiagnostic(input_path, output_directory_name = getwd())
}
